Split the oauth mod into three files.
