+++ Cargo.toml
@@ -21,3 +21,4 @@ url = "*"
 postgres = "*"
 r2d2 = "*"
 r2d2_postgres = "*"
+plugin = "*"
